tests/wpt/web-platform-tests/workers/modules/dedicated-worker-import-meta.html

<!DOCTYPE html>
<title>DedicatedWorker: import.meta</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

promise_test(() => {
  const script_url = 'resources/import-meta-url-worker.js';
  const worker = new Worker(script_url, { type: 'module' });
  return new Promise(resolve => worker.onmessage = resolve)
      .then(msg_event => assert_true(msg_event.data.endsWith(script_url)));
}, 'Test import.meta.url on the top-level module script.');

promise_test(() => {
  const script_url = 'import-meta-url-worker.js';
  const worker = new Worker('resources/dynamic-import-given-url-worker.js',
                            { type: 'module' });
  worker.postMessage('./' + script_url);
  return new Promise(resolve => worker.onmessage = resolve)
      .then(msg_event => assert_true(msg_event.data.endsWith(script_url)));
}, 'Test import.meta.url on the imported module script.');

</script>
